requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
instagrapi==2.0.0
instaloader==4.9.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.0
pandas==1.5.3
Pillow==10.0.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.9
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.2
six==1.16.0
socketify==0.0.27
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
